refactor(article): set base package of rdb adapter to "nettee.blolet.article.rdb"
