+++ services/article/driven/rdb/src/main/java/nettee/blolet/article/rdb/entity/ArticleLikesEntity.java
@@ -1,4 +1,4 @@
-package nettee.article.driven.rdb.entity;
+package nettee.blolet.article.rdb.entity;
 
 import jakarta.persistence.Convert;
 import jakarta.persistence.Entity;
@@ -6,8 +6,8 @@ import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
 import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
-import nettee.article.driven.rdb.entity.type.builder.ArticleLikesEntityStatus;
-import nettee.article.driven.rdb.entity.type.builder.ArticleLikesEntityStatusConverter;
+import nettee.blolet.article.rdb.entity.type.builder.ArticleLikesEntityStatus;
+import nettee.blolet.article.rdb.entity.type.builder.ArticleLikesEntityStatusConverter;
 import nettee.jpa.support.SnowflakeBaseTimeEntity;
 
 @SuperBuilder
