+++ services/article/driven/rdb/src/main/java/nettee/blolet/article/rdb/entity/type/builder/ArticleEntityStatus.java
@@ -1,6 +1,5 @@
-package nettee.article.driven.rdb.entity.type.builder;
+package nettee.blolet.article.rdb.entity.type.builder;
 
-import nettee.blolet.article.domain.sub.ArticleLikesStatus;
 import nettee.blolet.article.domain.sub.ArticleStatus;
 import nettee.common.marker.TypeSafeMarker.Present;
 import nettee.common.status.StatusCodeUtil;
@@ -13,8 +12,9 @@ import java.util.stream.Collectors;
 import static nettee.blolet.article.exception.ArticleErrorCode.DEFAULT;
 import static nettee.common.status.StatusParameters.GeneralPurposeFeatures.ALL;
 import static nettee.common.status.StatusParameters.GeneralPurposeFeatures.SUBITEM_READ;
+import static nettee.common.status.StatusParameters.GeneralPurposeFeatures.UPDATE;
 
-public enum ArticleLikesEntityStatus {
+public enum ArticleEntityStatus {
     REMOVED(StatusParameters.generate()
             .generalPurposeFeatures(SUBITEM_READ)
             .categoryBits(0b0000_0000_0000_0000)
@@ -31,46 +31,48 @@ public enum ArticleLikesEntityStatus {
             .instanceBits(0)
     ),
     SUSPENDED(StatusParameters.generate()
+            .generalPurposeFeatures(UPDATE, SUBITEM_READ)
             .categoryBits(0b0000_0000_0000_0100)
             .instanceBits(0)
-    );
+    ),
+    ;
 
     private final int code;
 
     static {
         assert Arrays.stream(values())
-                .map(ArticleLikesEntityStatus::getCode)
+                .map(ArticleEntityStatus::getCode)
                 .collect(Collectors.toSet())
                 .size()
                 == values().length
                 : "ArticleEntityStatus의 모든 code 필드가 고유해야 합니다.";
     }
 
-    ArticleLikesEntityStatus(StatusParameters<Present, Present> statusParameters) {
+    ArticleEntityStatus(StatusParameters<Present, Present> statusParameters) {
         this(statusParameters.encode(StatusCodeUtil::encode));
     }
 
-    ArticleLikesEntityStatus(int code) {
+    ArticleEntityStatus(int code) {
         this.code = code;
     }
 
     public int getCode() { return code; }
 
-    public static ArticleLikesEntityStatus valueOf(ArticleLikesStatus articleLikesStatus) {
+    public static ArticleEntityStatus valueOf(ArticleStatus articleStatus) {
         assert Set.of(ArticleStatus.REMOVED, ArticleStatus.PENDING, ArticleStatus.ACTIVE, ArticleStatus.SUSPENDED)
                 .containsAll(Arrays.stream(ArticleStatus.values()).collect(Collectors.toSet()))
-                : "ArticleLikesStatus 중 일부가 ArticleLikesEntityStatus::valueOf 함수에서 매핑되지 않습니다.";
+                : "ArticleStatus 중 일부가 ArticleEntityStatus::valueOf 함수에서 매핑되지 않습니다.";
 
-        return switch (articleLikesStatus) {
+        return switch (articleStatus) {
             case REMOVED -> REMOVED;
             case PENDING -> PENDING;
             case ACTIVE -> ACTIVE;
             case SUSPENDED -> SUSPENDED;
-            default -> throw new Error("ArticleLikesStatus 중 일부가 ArticleLikesEntityStatus::valueOf 함수에서 매핑되지 않습니다.");
+            default -> throw new Error("ArticleStatus 중 일부가 ArticleEntityStatus::valueOf 함수에서 매핑되지 않습니다.");
         };
     }
 
-    public static ArticleLikesEntityStatus valueOf(int value) {
+    public static ArticleEntityStatus valueOf(int value) {
         int categoryInstanceBits = 0xFFFFFF & value;
 
         return switch (categoryInstanceBits) {
