+++ services/article/driven/rdb/src/main/java/nettee/blolet/article/rdb/persistence/ArticleQueryAdapter.java
@@ -1,8 +1,8 @@
-package nettee.article.driven.rdb.persistence;
+package nettee.blolet.article.rdb.persistence;
 
 import lombok.RequiredArgsConstructor;
-import nettee.article.driven.rdb.persistence.mapper.ArticleEntityMapper;
 import nettee.blolet.article.application.port.ArticleQueryRepositoryPort;
+import nettee.blolet.article.rdb.persistence.mapper.ArticleEntityMapper;
 import nettee.blolet.article.readmodel.ArticleQueryModels.ArticleDetail;
 import nettee.blolet.article.readmodel.ArticleQueryModels.ArticleSummary;
 import org.springframework.data.domain.PageRequest;
