+++ services/article/driven/rdb/src/main/java/nettee/blolet/article/rdb/persistence/mapper/ArticleEntityMapper.java
@@ -1,9 +1,9 @@
-package nettee.article.driven.rdb.persistence.mapper;
+package nettee.blolet.article.rdb.persistence.mapper;
 
 import nettee.blolet.article.domain.Article;
-import nettee.article.driven.rdb.entity.ArticleEntity;
-import nettee.article.driven.rdb.projection.ArticleQueryProjections.ArticleDetailProjection;
-import nettee.article.driven.rdb.projection.ArticleQueryProjections.ArticleSummaryProjection;
+import nettee.blolet.article.rdb.entity.ArticleEntity;
+import nettee.blolet.article.rdb.projection.ArticleQueryProjections.ArticleDetailProjection;
+import nettee.blolet.article.rdb.projection.ArticleQueryProjections.ArticleSummaryProjection;
 import nettee.blolet.article.readmodel.ArticleQueryModels.ArticleDetail;
 import nettee.blolet.article.readmodel.ArticleQueryModels.ArticleSummary;
 import org.mapstruct.Mapper;
